# perf(history): enable sql index on chat_id in history table