This makes every query with a "WHERE history.chat_id" clause quicker,
improving history load time by 50% on my profile.

This PR is based on top of https://github.com/qTox/qTox/pull/5872.

- [x] Commits follow our [git commit guidelines](https://github.com/qTox/qTox/blob/master/CONTRIBUTING.md#git-commit-guidelines)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/qtox/qtox/5873)
<!-- Reviewable:end -->