@anthonybilinski 2019-10-12
I checked how much one of my profiles database grew by on update, since storing the index adds data to the database. The change I saw was:
version 3: 2224128
version 4: 2256896
with a `history` table with 3348 entries.

So a little over a 1% increase in size. This seems well worth it for the ~50% start up time savings.