@anthonybilinski 2019-10-13
And for the actual performance numbers, on my main profile with.. a lot of history, I timed in rawdatabase the total amount of time spent inside sql_step from when qTox is opened until when the GUI appears.
Before change:
1st run: 16273 ms
2nd run: 16389 ms
After change:
1st run: 8712 ms
2nd run: 8637 ms

This still seems unreasonably long to open the program, but the other points mentioned in https://github.com/qTox/qTox/issues/5812 can continue to improve that.