@anthonybilinski 2019-10-13
No it doesn't, the only relation to #5872 is that both increment the schema_version in our upgrade code. We can definitely merge the other and let is sit for a bit before merging this one.