@anthonybilinski 2019-11-05
https://github.com/qTox/qTox/pull/5872 is now merged, so this is the next db upgrade in the queue. I'll hold of merging it for a bit to make sure https://github.com/qTox/qTox/pull/5872 settles well, but feel free to review